requirements.txt
openai-agents-sdk
python-dotenv
fastapi
uvicorn
httpx
beautifulsoup4
redis
